implemented "ask your followers"

+++ app/models/inbox.rb
@@ -12,10 +12,10 @@ class Inbox < ActiveRecord::Base
 
   def remove
     unless self.question.user.nil?
-      self.question.user.decrement! :asked_count
+      self.question.user.decrement! :asked_count if self.question.answers.count == 1
     end
 
-    self.question.destroy
+    self.question.destroy if self.question.answers.count == 1
     self.destroy
   end
 end
